cancel_timer_2 async true
