+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test/with_reference_timer_reference/with_list_options.rs
@@ -161,32 +161,6 @@ fn with_info_false_without_timer_returns_ok(options: fn(&Process) -> Term) {
     });
 }
 
-fn with_timeout_returns_ok_after_timeout_message_was_sent(options: fn(&Process) -> Term) {
-    with_timer_in_different_thread(|milliseconds, barrier, timer_reference, process| {
-        timeout_after_half_and_wait(milliseconds, barrier);
-        timeout_after_half_and_wait(milliseconds, barrier);
-
-        let timeout_message = different_timeout_message(timer_reference, process);
-
-        assert_eq!(receive_message(process), Some(timeout_message));
-
-        let cancel_timer_message = cancel_timer_message(timer_reference, false.into(), process);
-
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-        assert_eq!(receive_message(process), Some(cancel_timer_message));
-
-        // again
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-        assert_eq!(receive_message(process), Some(cancel_timer_message));
-    });
-}
-
 fn without_info_without_local_reference_errors_badarg(
     source_file: &'static str,
     options: fn(&Process) -> Term,
