+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test/with_reference_timer_reference/with_list_options/with_async_true/without_info/with_local_reference/with_timer/in_same_thread.rs
@@ -1,85 +1,14 @@
 use super::*;
 
-use crate::test::{timeout_after_half, with_timer_in_same_thread};
-
 #[test]
 #[ignore]
 fn without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message() {
-    with_timer_in_same_thread(|milliseconds, message, timer_reference, process| {
-        timeout_after_half(milliseconds);
-
-        let timeout_message = timeout_message(timer_reference, message, process);
-
-        assert!(!has_message(process, timeout_message));
-
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-
-        let received_message = receive_message(process).unwrap();
-
-        let received_tuple_result: core::result::Result<Boxed<Tuple>, _> =
-            received_message.try_into();
-
-        assert!(received_tuple_result.is_ok());
-
-        let received_tuple = received_tuple_result.unwrap();
-
-        assert_eq!(received_tuple[0], Atom::str_to_term("cancel_timer"));
-        assert_eq!(received_tuple[1], timer_reference);
-
-        let milliseconds_remaining = received_tuple[2];
-
-        assert!(milliseconds_remaining.is_integer());
-        assert!(process.integer(0).unwrap() < milliseconds_remaining);
-        assert!(milliseconds_remaining <= process.integer(milliseconds / 2).unwrap());
-
-        let false_cancel_timer_message =
-            cancel_timer_message(timer_reference, false.into(), process);
-
-        // again before timeout
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-        assert_eq!(receive_message(process), Some(false_cancel_timer_message));
-
-        timeout_after_half(milliseconds);
-
-        assert!(!has_message(process, timeout_message));
-
-        // again after timeout
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-        assert_eq!(receive_message(process), Some(false_cancel_timer_message));
-    })
+    super::without_timeout_returns_milliseconds_remaining_and_does_not_send_timeout_message(
+        options,
+    );
 }
 
 #[test]
 fn with_timeout_returns_ok_after_timeout_message_was_sent() {
-    with_timer_in_same_thread(|milliseconds, message, timer_reference, process| {
-        timeout_after(milliseconds);
-
-        let timeout_message = timeout_message(timer_reference, message, process);
-
-        assert_eq!(receive_message(process), Some(timeout_message));
-
-        let cancel_timer_message = cancel_timer_message(timer_reference, false.into(), process);
-
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-        assert_eq!(receive_message(process), Some(cancel_timer_message));
-
-        // again
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-        assert_eq!(receive_message(process), Some(cancel_timer_message));
-    })
+    super::with_timeout_returns_ok_after_timeout_message_was_sent(options);
 }
